requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=0.0.6
google-cloud-aiplatform>=1.38.0

# Database
supabase>=2.0.0

# Search
tavily-python>=0.3.0

# Data ingestion
apify-client>=1.6.0
feedparser>=6.0.10

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
